+++ lib/app_links.dart
@@ -3,8 +3,10 @@ import 'dart:async';
 import 'package:flutter/services.dart';
 
 /// Callback when your app is woke up by an incoming link
+/// [uri] and [stringUri] are same value.
+/// [stringUri] is available for custom handling like uppercased uri.
 ///
-typedef void OnAppLinkFunction(Uri uri);
+typedef void OnAppLinkFunction(Uri uri, String stringUri);
 
 /// Android App Links, Deep Links,
 /// iOs Universal Links and Custom URL schemes handler
@@ -34,7 +36,8 @@ class AppLinks {
         switch (call.method) {
           case _onAppLinkMethod:
             if (call.arguments != null) {
-              onAppLink(Uri.parse(call.arguments.toString()));
+              final uri = call.arguments.toString();
+              onAppLink(Uri.parse(uri), uri);
             }
         }
 
@@ -46,18 +49,30 @@ class AppLinks {
   /// Gets the initial / first link
   /// returns [Uri] or [null]
   Future<Uri?> getInitialAppLink() async {
-    final result = await _channel.invokeMethod(_getInitialAppLinkMethod);
+    final result = await getInitialAppLinkString();
+    if (result == null) return null;
+
+    return Uri.tryParse(result);
+  }
+
+  /// Gets the initial / first link
+  /// returns [Uri] as String or [null]
+  Future<String?> getInitialAppLinkString() async {
+    return await _channel.invokeMethod(_getInitialAppLinkMethod);
+  }
+
+  /// Gets the latest link
+  /// returns [Uri] or [null]
+  Future<Uri?> getLatestAppLink() async {
+    final result = await getLatestAppLinkString();
     if (result == null) return null;
 
     return Uri.tryParse(result);
   }
 
   /// Gets the latest link
-  /// returns [Uri] or [null]
-  Future<Uri?> getLatestAppLink() async {
-    final result = await _channel.invokeMethod(_getLatestAppLinkMethod);
-    if (result == null) return null;
-
-    return Uri.tryParse(result);
+  /// returns [Uri] as String or [null]
+  Future<String?> getLatestAppLinkString() async {
+    return await _channel.invokeMethod(_getLatestAppLinkMethod);
   }
 }